# Contraction sequence optimization