--- a/src/ContractionSequenceOptimization/ContractionSequenceOptimization.jl
+++ b/src/ContractionSequenceOptimization/ContractionSequenceOptimization.jl
@@ -1,7 +1,12 @@
 
 module ContractionSequenceOptimization
 
-  export contraction_sequence
+  using NDTensors
+
+  import NDTensors:
+    dim
+
+  export optimized_contraction_sequence
 
   include("utils.jl")
   include("three_tensors.jl")
@@ -12,17 +17,17 @@ module ContractionSequenceOptimization
   const DimT = UInt128
 
   """
-      contraction_sequence(T)
+      optimized_contraction_sequence(T)
 
   Returns a contraction sequence for contracting the tensors `T`. The sequence is generally optimal (currently, outer product contractions are skipped, but some optimal sequences require outer product contractions).
   """
-  function contraction_sequence(T)
+  function optimized_contraction_sequence(T)
     if length(T) == 1
       return Any[1]
     elseif length(T) == 2
       return Any[1, 2]
     elseif length(T) == 3
-      return contraction_sequence(T[1], T[2], T[3])
+      return optimized_contraction_sequence(T[1], T[2], T[3])
     end
     return breadth_first_constructive(T)
   end