UPDATED

This PR adds an optional mode to ITensors.jl to automatically find an optimal contraction sequence for a set of tensors. Here are some examples with timings:
```julia
using ITensors
using BenchmarkTools
using LinearAlgebra

BLAS.set_num_threads(2)

i = Index(100, "i")
A = randomITensor(i', dag(i))

ITensors.enable_contraction_sequence_optimization()

#
# 2 tensors
#

println("\n2 tensors")

# Simple pair contraction
@btime $A' * $A

#
# 3 tensors
#

println("\n3 tensors")

println("Pick sequence explicitly")
@btime $A'' * ($A' * $A)

println("Let it optimize")
@btime $A'' * $A' * $A
@btime $A'' * $A * $A'

#
# 4 tensors
#

println("\n4 tensors")

println("Pick sequence explicitly")
@btime $A''' * ($A'' * ($A' * $A))

println("Let it optimize")
@btime $A''' * $A'' * $A' * $A
@btime $A'' * $A * $A''' * $A'

#
# 5 tensors
#

println("\n5 tensors")

println("Pick sequence explicitly")
@btime $A'''' * ($A''' * ($A'' * ($A' * $A)))

println("Let it optimize")
@btime $A'''' * $A''' * $A'' * $A' * $A
@btime $A'' * $A'''' * $A * $A''' * $A'

ITensors.disable_contraction_sequence_optimization()
```
with results:
```julia
2 tensors
  20.118 μs (26 allocations: 80.72 KiB)

3 tensors
Pick sequence explicitly
  42.103 μs (55 allocations: 161.91 KiB)
Let it optimize
  45.886 μs (62 allocations: 162.25 KiB)
  47.855 μs (62 allocations: 162.25 KiB)

4 tensors
Pick sequence explicitly
  64.827 μs (87 allocations: 243.56 KiB)
Let it optimize
  73.744 μs (197 allocations: 254.22 KiB)
  82.323 μs (198 allocations: 254.27 KiB)

5 tensors
Pick sequence explicitly
  87.252 μs (122 allocations: 325.69 KiB)
Let it optimize
  103.298 μs (278 allocations: 340.11 KiB)
  103.221 μs (279 allocations: 340.16 KiB)
```
So you can see the contraction sequence optimization has an overhead of a few microseconds (increasing with the number of tensors).